Revert "Revert "Solicit email consents using new Identity Consent API""
