+++ app/views/giraffe/postPayment.scala.html
@@ -14,9 +14,10 @@
         <div class="contribute-section postPaymentContact">
             <div class ="circleBackground postPaymentEnvelope"></div>
             <p class ="contribute-form__title">We would like to stay in touch</p>
-            <form action="update-metadata" method="POST" >
+            <form action="update-marketing" method="POST" >
                 <div class="giraffe-checkbox">
-                    <input id="marketingOptIn" type="checkbox" name="marketingOptIn" checked="true" value="true"/>
+                    <input id="marketingOptIn" type="checkbox" name="marketingOptIn" value="true"/>
+                    <!--THIS COPY NEEDS TO CHANGE BEFORE MERGING -->
                     <label for="marketingOptIn">Keep me informed with updates and offers from the Guardian</label>
                 </div>
                 <button class="action action--button action--button--forward action action--button contribute-navigation__next action--next">
