+++ assets/javascripts/src/components/pages/Details.jsx
@@ -20,13 +20,6 @@ export default class Details extends React.Component {
                         value={this.props.details.postcode}
                         onChange={ event => this.props.updateDetails({ postcode: event.target.value })}
                         />
-
-            <div className="giraffe-checkbox">
-                <input id="guardian-opt-in" type="checkbox" name="guardian-opt-in"
-                       checked={this.props.details.optIn}
-                       onChange={event => this.props.updateDetails({ optIn: event.target.checked })} />
-                <label htmlFor="guardian-opt-in">Keep me up to date with offers from the Guardian</label>
-            </div>
         </div>
     }
 }
