+++ build.sbt
@@ -60,7 +60,8 @@ val paypalSdk = "com.paypal.sdk" % "rest-api-sdk" % "1.13.0" exclude("org.apache
 val macwire = "com.softwaremill.macwire" %% "macros" % "2.2.2" % Provided
 val anormLib = "com.typesafe.play" %% "anorm" % "2.5.2"
 val postgresql = "org.postgresql" % "postgresql" % "9.4.1209"
-val identityCookie =  "com.gu.identity" %% "identity-cookie" % "3.51"
+val identityCookie =  "com.gu.identity" %% "identity-cookie" % "3.100"
+val identityModel =  "com.gu.identity" %% "identity-model" % "3.100"
 val scalaTest = "org.scalatestplus.play" %% "scalatestplus-play" % "2.0.0"
 val enumeratum = "com.beachape" %% "enumeratum" % "1.5.12"
 val enumeratumJson = "com.beachape" %% "enumeratum-play-json" % "1.5.12"
